更改变量结构 vmat，现在 vmat 类型支持大量数据快速查询

+++ BaseFunction.h
@@ -4,8 +4,8 @@
 typedef struct Interface_Base {
 	// string.h
 	u32	(*vname_check)		(char *name);
-	u64	(*vname_gen)		(unsigned char *name);
-	u64	(*vhead_gen)		(u64 head);
+	u32	(*vname_gen)		(unsigned char *name);
+	u32	(*vhead_gen)		(u64 head);
 	u64	(*vname_head)		(unsigned char *name);
 	char*	(*skip_string)		(char *exp);
 	char*	(*get_string)		(char *exp, char **pexp);
